+++ spotbugsTestCases/src/java/BadUnsafeSun.java
@@ -1,0 +1,13 @@
+import java.lang.reflect.Field;
+import sun.misc.Unsafe;
+
+class BadUnsafeSun {
+
+  void method() throws Exception {
+    final Field f = Unsafe.class.getDeclaredField("theUnsafe");
+    f.setAccessible(true);
+    final Unsafe unsafe = (Unsafe) f.get(null);
+    unsafe.putAddress(0, 0);
+  }
+
+}